- Modification's bonus click percentage implemented - Numbers output changed to double type

+++ Models/Upgrade.cs
@@ -11,7 +11,7 @@
 
         public void IncreaseCost()
         {
-            Cost = (int)(Cost * CostMultiplier);
+            Cost = (Cost * CostMultiplier);
         }
     }
 }
